撤回 1.5x、闲聊文案进气泡池、更正“陪”为“赔”
- 显示倍率回到整数。1.5x 在锐利度上被整数档完全支配：实际生效的是
  项目倍率 x 系统 DPR，1.5x 只在 200% 系统缩放下得到整数物理倍率，
  而那一档 1x 与 2x 本来就满足。BubbleRenderer 的倍率收回 int，
  布局测试三档改为 1x / 2x / 3x。
- 第 4.1 节补一条：随机气泡允许选中多页文案，按连续对话规则显示，
  不套用单句气泡的自动消失时长。四条两页闲聊文案据此进入随机池。
- 第 4.4 节冰淇淋掉落第 4 页“你要陪我”更正为“你要赔我”。上下文是
  “我已经没有钱买第二个了”，指赔偿。属转录错字更正，不是改写原作台词。

18 个测试目标全过，编译零警告。

+++ src/ui/BubbleRenderer.h
@@ -20,11 +20,12 @@ namespace mub::ui {
 class BubbleRenderer
 {
 public:
-    explicit BubbleRenderer(double scale = 1.0);
+    explicit BubbleRenderer(int scale = 1);
 
-    // 显示倍率。第 5.1 节：`1×`、`1.5×`、`2×` 等离散档位，一律最近邻采样。
-    void setScale(double scale);
-    double scale() const;
+    // 显示倍率。第 5.1 节：只用整数倍率，一律最近邻采样。
+    // 与角色窗口共用同一个倍率，气泡和角色因此始终等比。
+    void setScale(int scale);
+    int scale() const;
 
     // `face` 为空表示该页没有可用表情素材，此时只画文字区。
     // 正常内容里每页都有表情（第 4 节），空表情属于素材缺失的降级路径。
@@ -59,7 +60,7 @@ private:
     int textAreaWidth() const;
     QStringList computeWrappedLines() const;
 
-    double scale_ = 1.0;
+    int scale_ = 1;
     QImage face_;
     QString fullText_;
     int visibleCharacters_ = 0;
